requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2